requirements.txt
discord.py
python-dotenv
black